+++ lib/main.dart
@@ -15,6 +15,7 @@ import 'package:nota_note/pages/on_boarding_page/on_boarding_page.dart';
 import 'package:flutter_quill/flutter_quill.dart';
 import 'package:flutter_localizations/flutter_localizations.dart';
 import 'services/local_storage_service.dart';
+import 'pages/my_home_page/my_home_page.dart';
 
 import 'package:timeago/timeago.dart' as timeago;
 
@@ -86,89 +87,12 @@ class MyApp extends ConsumerWidget {
                 Scaffold(body: Center(child: Text('유저 정보 에러: $e'))),
             data: (user) {
               if (user == null) return const LoginPage();
-              return const MyHomePage();
+              return const OnBoardingPage();
+              //return const MyHomePage(); 테스트용 홈페이지로 복귀하고 싶을 때 돌려놓으세요
+
             },
           );
         },
-      ),
-    );
-  }
-}
-
-class MyHomePage extends StatelessWidget {
-  const MyHomePage({super.key});
-
-  @override
-  Widget build(BuildContext context) {
-    return Scaffold(
-      body: Center(
-        child: Column(
-          mainAxisAlignment: MainAxisAlignment.center,
-          children: [
-            const Text('NotaNote 예시 페이지'),
-            const SizedBox(height: 20),
-            ElevatedButton(
-              onPressed: () {
-                // MainPage로 이동
-                Navigator.push(
-                  context,
-                  MaterialPageRoute(builder: (context) => const MainPage()),
-                );
-              },
-              child: const Text('메인 페이지로 이동'),
-            ),
-            ElevatedButton(
-              onPressed: () {
-                // 테스트 데이터로 MemoGroupPage 직접 이동 (이전 코드 용도)
-                Navigator.push(
-                  context,
-                  MaterialPageRoute(
-                      builder: (context) => const MemoGroupPage(
-                            groupId: 'group1',
-                            groupName: '그룹1',
-                          )),
-                );
-              },
-              child: const Text('메모 그룹 페이지로 이동 (테스트)'),
-            ),
-            ElevatedButton(
-              onPressed: () {
-                Navigator.push(
-                  context,
-                  MaterialPageRoute(
-                    builder: (context) => MemoPage(
-                      groupId: 'group1',
-                      noteId: 'note1',
-                      pageId: 'page1',
-                    ),
-                  ),
-                );
-              },
-              child: const Text('메모 페이지로 이동'),
-            ),
-            ElevatedButton(
-              onPressed: () {
-                Navigator.push(
-                  context,
-                  MaterialPageRoute(
-                      builder: (context) => const OnBoardingPage()),
-                );
-              },
-              child: const Text('온보딩 페이지로 이동'),
-            ),
-            ElevatedButton(
-              onPressed: () async {
-                await FirebaseAuth.instance.signOut();
-                if (!context.mounted) return;
-                Navigator.pushReplacement(
-                  context,
-                  MaterialPageRoute(builder: (_) => const LoginPage()),
-                );
-              },
-              child: const Text('로그아웃'),
-            ),
-          ],
-        ),
       ),
     );
   }
